LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-showModal-expected.txt


PASS dialog element: showModal()
PASS showModal() on a <dialog> that already has an open attribute throws an InvalidStateError exception
PASS showModal() on a <dialog> after initial showModal() and removing the open attribute
PASS showModal() on a <dialog> not in a Document throws an InvalidStateError exception
PASS when opening multiple dialogs, only the newest one is non-inert
FAIL opening dialog without focusable children assert_equals: expected Element node <dialog id="d6" open=""></dialog> but got Element node <body><div id="log"></div>
<button id="b0">OK</button>
<d...
PASS opening dialog with multiple focusable children
PASS opening dialog with multiple focusable children, one having the autofocus attribute
PASS when opening multiple dialogs, the most recently opened is rendered on top
PASS Although the document is not attached to any pages, showModal() should execute as normal.
OK
